+++ frontend/src/app/infrastructure/adapters/secondary/beat-source/json-files-reader.service.ts
@@ -9,7 +9,14 @@ export class JsonFileReader implements JsonFilesReaderInterface {
   constructor(private readonly http: HttpClient) {}
 
   loadAllJson(): Observable<CompactBeat[]> {
-    const files = ['metal-metal.json', 'metal-metal2.json'];
+    let files = ['techno/techno.json', 'techno/off-beat-clap.json'];
+    files = files.concat('metal/metal.json', 'metal/metal-blastbeat.json', 'metal/half-time-groove.json');
+    files = files.concat('rock/rock.json', 'rock/variation.json');
+    files = files.concat('punk/punk-beat-quarter-note-groove.json', 'punk/punk-beat-quarter-note-groove-variation.json', 'punk/punk-beat-eight-note-fill.json');
+    files = files.concat('psytrance/psytrance.json');
+    files = files.concat('dancehall/standard.json');
+    files = files.concat('techno-hardcore/gabber.json');
+    files = files.concat('dub/dub.json');
 
     const requests = files
       .map(file => this.http.get('assets/beats/' + file))
