+++ frontend/src/assets/beats/dancehall/standard.json
@@ -1,0 +1,22 @@
+{
+  "label": "Standard dancehall",
+  "genre": "Dancehall",
+  "bpm": 105,
+  "tracks": [
+    {
+      "name": "Hats",
+      "fileName": "techno/hat.wav",
+      "steps": "XXXXXXXXXXXXXXXX"
+    },
+    {
+      "name": "Snare",
+      "fileName": "techno/snare.wav",
+      "steps": "      X       X "
+    },
+    {
+      "name": "Kick",
+      "fileName": "techno/kick.wav",
+      "steps": "X  X  X X  X  X "
+    }
+  ]
+}
