+++ frontend/src/assets/beats/metal/half-time-groove.json
@@ -1,0 +1,22 @@
+{
+  "label": "Half time groove",
+  "genre": "Metal",
+  "bpm": 145,
+  "tracks": [
+    {
+      "name": "Snare",
+      "fileName": "metal/snare.mp3",
+      "steps": "        X       "
+    },
+    {
+      "name": "Crash",
+      "fileName": "metal/crash.mp3",
+      "steps": "X   X   X   X   "
+    },
+    {
+      "name": "Kick",
+      "fileName": "metal/kick.mp3",
+      "steps": "X               "
+    }
+  ]
+}
