+++ frontend/src/assets/beats/metal/metal.json
