+++ frontend/src/assets/beats/rock/rock.json
@@ -1,0 +1,22 @@
+{
+  "label": "Rock",
+  "genre": "Rock",
+  "bpm": 145,
+  "tracks": [
+    {
+      "name": "Snare",
+      "fileName": "metal/snare.mp3",
+      "steps": "    X       X   "
+    },
+    {
+      "name": "Hats",
+      "fileName": "metal/hat.mp3",
+      "steps": "X   X   X   X   "
+    },
+    {
+      "name": "Kick",
+      "fileName": "metal/kick.mp3",
+      "steps": "X X       X     "
+    }
+  ]
+}
