+++ frontend/src/assets/beats/rock/variation.json
@@ -1,0 +1,22 @@
+{
+  "label": "Rock variation",
+  "genre": "Rock",
+  "bpm": 145,
+  "tracks": [
+    {
+      "name": "Snare",
+      "fileName": "metal/snare.mp3",
+      "steps": "    X       X   "
+    },
+    {
+      "name": "Crash",
+      "fileName": "metal/crash.mp3",
+      "steps": "X   X   X   X   "
+    },
+    {
+      "name": "Kick",
+      "fileName": "metal/kick.mp3",
+      "steps": "X       X       "
+    }
+  ]
+}
