+++ frontend/src/assets/beats/techno-hardcore/gabber.json
@@ -1,0 +1,27 @@
+{
+  "label": "Gabber",
+  "genre": "Techno Hardcore",
+  "bpm": 200,
+  "tracks": [
+    {
+      "name": "Crash Cymbal",
+      "fileName": "gabber/crash.wav",
+      "steps": "                               X"
+    },
+    {
+      "name": "Open HiHats",
+      "fileName": "gabber/open-hihat.wav",
+      "steps": "X X X X X X X X X X X X X X X X "
+    },
+    {
+      "name": "Clap",
+      "fileName": "gabber/clap.wav",
+      "steps": "X   X   X   X   X   X   X   X   "
+    },
+    {
+      "name": "Kick",
+      "fileName": "gabber/kick.wav",
+      "steps": "X   X   X   X   X   X   X   X   "
+    }
+  ]
+}
