+++ frontend/src/assets/beats/techno/off-beat-clap.json
@@ -1,0 +1,22 @@
+{
+  "label": "OffBeat clap",
+  "genre": "Techno",
+  "bpm": 128,
+  "tracks": [
+    {
+      "name": "Clap",
+      "fileName": "gabber/clap.wav",
+      "steps": "    X      XX       X      XX   "
+    },
+    {
+      "name": "Hats",
+      "fileName": "techno/hat.wav",
+      "steps": "  X   X   X   X   X   X   X   X "
+    },
+    {
+      "name": "Kick",
+      "fileName": "techno/kick.wav",
+      "steps": "X   X   X   X   X   X   X   X   "
+    }
+  ]
+}
